Extract ConfigurationCacheSupportedTypesIntegrationTest
from ConfigurationCacheIntegrationTest
